jit: Add new provisioning handler
Skip system role deletion on default provisionning handler might
conflit with IAM scenarios. Therefore introduced a new handler
shipped by default, which will skip deleting system roles
